Added some corrections and tests.

+++ api/src/main/java/ar/edu/unaj/api/beans/domain/Address.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class Address {
 
     private String street;
-    private long number;
+    private Integer number;
     private String city;
     private String state;
     private String country;
